+++ CHANGELOG.md
@@ -8,9 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming] - 2020-01-08
 
 - Add `volume` and `isMuted` prop to `Song`
-- Add `polyphony` and `oscillatorType` prop to `Instrument`
+- Add `polyphony`, `oscillatorType`, `envelopeAttack`, `envelopeDecay`, `envelopeSustain` and `envelopeRelease` props to `Instrument`
 - Remove `type` of `polySynth` from `Instrument` as it is just a wrapper around other synths
-- Add polyphony for Instrument synths
 - Add `membraneSynth`, `metalSynth` and `pluckSynth` types to `Instrument`
 - Update `AMSynth` and `FMSynth` values in `Instrument` `type` to `amSynth` and `fmSynth`
 - Add more Typescript definitions