+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Upcoming] - 2020-01-08
 
+- Change `tempo` to `bpm` in `Song` to match `Tone` API
 - Add `volume` and `isMuted` prop to `Song`
 - Add `polyphony`, `oscillatorType`, `envelopeAttack`, `envelopeDecay`, `envelopeSustain` and `envelopeRelease` props to `Instrument`
 - Remove `type` of `polySynth` from `Instrument` as it is just a wrapper around other synths
 - Add `membraneSynth`, `metalSynth` and `pluckSynth` types to `Instrument`
 - Update `AMSynth` and `FMSynth` values in `Instrument` `type` to `amSynth` and `fmSynth`
+- Remove `/example` folder to focus on `/website` that already includes docs and examples
 - Add more Typescript definitions
 
 ## [0.2.2] - 2019-09-18