+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `/example` folder to focus on `/website` that already includes docs and examples
 - Change `steps` from `step.note` to `step.name` in `Track`
 - Add more Typescript definitions
+- Change `Instrument` to use `useLayoutEffect` instead of `useEffect` for triggering `notes`
+- Ensure `Instrument` and `Effect` don't crash if unknown `type` is passed
 
 ## [0.2.2] - 2019-09-18