+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   <!-- - Change `Instrument` to use `useLayoutEffect` instead of `useEffect` for triggering `notes` -->
 - Ensure `Instrument` and `Effect` don't crash if unknown `type` is passed
 - Update `onStepPlay` arguments to `StepNoteType[]` and `number`.
+- Add `wet` prop to `Effect`
 
 ## [0.2.2] - 2019-09-18