+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ensure `Instrument` and `Effect` don't crash if unknown `type` is passed
 - Update `onStepPlay` arguments to `StepNoteType[]` and `number`.
 - Add `wet` prop to `Effect`
+- Update `propTypes`
+- Update docs
 
 ## [0.2.2] - 2019-09-18