+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Upcoming] - 2020-01-08
+## [0.3.0] - 2020-02-07
 
 - Change `tempo` to `bpm` in `Song` to match `Tone` API
 - Add `volume` and `isMuted` prop to `Song`